requirements.txt
# Core libraries for ML modeling and data wrangling
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
pyarrow==15.0.2
joblib==1.4.2

# Visualization & Data exploration
seaborn

# Experiment logging & model tracking
mlflow==2.12.1

# Workflow orchestration and automation
prefect==2.16.3
griffe==0.41.3
setuptools>=68.0.0


# Application and API framework
fastapi==0.111.0
uvicorn==0.30.1

# Monitoring & Data quality evaluation
evidently==0.4.17

# Configuration handling
python-dotenv==1.0.1



# Testing
pytest==8.2.1

# Development tools
black==24.4.2
flake8==7.0.0

# Database interaction
psycopg2-binary==2.9.9
boto3==1.40.0
